[+] fix : date and time picker refactor + reformat
